Seperated out code by implementing testing system.
Seperated scr/main.py into:
1) scr/getDbFilePath.py
The sofar-singularly-impemented function for getting the database's path.
2) tests/test_getFilePath.py
The test related to the above function, for inital testing of the testing system.
3) tests/RUNALLTESTS.sh
The bash file responsible for running & reporting all tests. Awful.

Note:
test_*.py files are to be improved by standardising (something along the lines of) a '@test' decorator in order to heavily reformat them, as the current design for the singular test is unmaintainable over larger projects. This will be done on the develop-standard-testing branch - at the time still to be created.