Hotfixed unnessesary code complexity and commenting - CORRECTLY THIS TIME
A comment stated that double-backlash (\\) had to be used in place of a singular backslash (\) within a filepath (which was represented as a string).
This was originally done to prevent the backslash being interpreted as an escape code, but it can instead be entierly curcumvented by using raw strings (r"filepathgoeshere").
Therefore, the string can be raw'ed and the comment removed.
This is the second attempted commit of this - the previous one was reverted for improper file placement. Whoops.